code-detective-model/requirements.txt
# run commmand pip install -r requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.31.0
protobuf==5.27.0
pydantic==2.7.3
pydantic_core==2.18.4
setuptools==70.0.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1